一个完整接口

+++ sql2api-manager/src/main/java/com/wkclz/sql2api/manager/dao/ApiGroupMapper.java
@@ -9,7 +9,7 @@ import java.util.List;
 public interface ApiGroupMapper {
 
     // 分页
-    Integer getApiGroupCount(ApiGroupDto dto);
+    Long getApiGroupCount(ApiGroupDto dto);
     List<ApiGroupDto> getApiGroupList(ApiGroupDto dto);
 
 }
